add: 이미지, delete: 불필요한 내용 삭제, move: 메인 코드 위치 변경
15x15 이미지 추가
개인 불필요한 파일 삭제
main의 commons 이동
