+++ Assets/Scripts/Commons/Singleton.cs
