+++ Assets/Scripts/Commons/Singleton.cs.meta
