--- Assets/workspace/YU__FFE/Scripts/willbedeleted.meta
@@ -1,3 +1,0 @@
-fileFormatVersion: 2
-guid: 1f793366bb3a4c2283173ae1b1ba2a82
-timeCreated: 1741747026
