--- Assets/workspace/YU__FFE/Scripts/willbedeleted/CrackTest1.meta
@@ -1,3 +1,0 @@
-fileFormatVersion: 2
-guid: b083de0f3bcb4711b2c7b51bfcc005e2
-timeCreated: 1741747040
