--- Assets/workspace/YU__FFE/Scripts/willbedeleted/CrackTest1/Crack1.cs
@@ -1,5 +1,0 @@
-namespace workspace.YU__FFE.Scripts.willbedeleted.CrackTest1 {
-    public class Crack1 {
-        
-    }
-}
