--- Assets/workspace/YU__FFE/Scripts/willbedeleted/CrackTest1/Crack1.cs.meta
@@ -1,3 +1,0 @@
-fileFormatVersion: 2
-guid: 7c67709a03954be7925b4c6a33783f5a
-timeCreated: 1741747045
