--- Assets/workspace/YU__FFE/Scripts/willbedeleted/CrackTest2.meta
@@ -1,3 +1,0 @@
-fileFormatVersion: 2
-guid: c34cbdb685654448a3d62ce7440e7aed
-timeCreated: 1741747049
