--- Assets/workspace/YU__FFE/Scripts/willbedeleted/CrackTest2/Crack1.cs
@@ -1,5 +1,0 @@
-namespace workspace.YU__FFE.Scripts.willbedeleted.CrackTest2 {
-    public class Crack1 {
-        
-    }
-}
