--- Assets/workspace/YU__FFE/Scripts/willbedeleted/CrackTest2/Crack1.cs.meta
@@ -1,3 +1,0 @@
-fileFormatVersion: 2
-guid: 0ac56a45f700483fa3f4aa34266d6354
-timeCreated: 1741747049
